css-writing-modes-3_dev/html/writing-mode-parsing-svg1-001.htm

<!DOCTYPE html>
<html><head><title>CSS Writing Modes: parsing SVG 1.1 writing-mode values</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii">
<link href="https://drafts.csswg.org/css-writing-modes-3/#svg-writing-mode" rel="help">
<meta content="This test asserts the parser, inherits, and getComputedStyle works correctly for the SVG 1.1 writing-mode values." name="assert">
<meta content="dom svg may" name="flags">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

</head><body><div data-expected="lr" style="writing-mode: lr" title="writing-mode: lr"></div>
<div data-expected="rl" style="writing-mode: rl" title="writing-mode: rl"></div>

<div data-expected="tb" style="writing-mode: tb" title="writing-mode: tb"></div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).writingMode;
    assert_equals(actual, element.dataset.expected);
  }, element.title);
});
</script>
</body></html>
